MB-20054: [BP] engine_testapp: Fully enable stderr logging in engine_testapp
engine_testapp has a -X flag to enable logging to stderr (by default
output is logged to null). This works for the instances where the
logger is passed directly to the client; for example during initial
engine load; however it doesn't work for the actual engine itself;
which will use the server APIs to obtain the logger from the
mock_server - which is always null.

Expand the support in mock_server / engine_testapp to allow '-X' to
also enable the stderr logger in mock_server for the use of the
engine. This vastly improves the information logged when running with
'-X'.

Change-Id: I96a3b1880e2e7de6214cdb47e6f484be05d49019
Reviewed-on: http://review.couchbase.org/65610
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>