Use default kvstore for vbstate for faster indexing
For single node case there are 256 kvstores (1 for each vbucket)
created in a single forestdb file. Now creating an additional
kvstore for the vbstate makes for a total of 257 kvstores.
In multi-kv mode, forestdb uses HB+trie for the sequence trees.
As only 256 kvstores can be represented in a single byte, two chunks
are needed to distinguish 257 kvstores. But by using the default
kvstore for the vbstate, we can keep the trie functioning within
the first chunk, thus speeding up index traversals on sequence numbers.

Change-Id: Ic2b589a54d74b96abea94a27c93742f900c56d34
Reviewed-on: http://review.couchbase.org/53255
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>