MB-15780: Fix instability in topkeys tests
Fix an issue with the topkeys tests where we occasionally got
incorrect results. The issue was in get_topkeys_json_value() where the
second read (of the terminating NULL packet) used the same recv buffer
as the first data packet, and moreover there was no null-termination
added to the received data.

Fix these issues and generally improve the robustness of the test.

Change-Id: I98784cc10954cc81c0530c6a6d4425cc927af2d4
Reviewed-on: http://review.couchbase.org/53484
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: buildbot <build@couchbase.com>