testapp: Uniquify the various per-process variables passed to memcached
Where unique files / port numbers need to be specified to memcached;
include the current time (in addition to testapp's process ID). We now
have some tests which restart memcached; and hence if only the pid of
testapp is used then there will collisions.

Note that we also need to record the original SSL port chosen to allow
the tests which reconfigure the running memcached instance to use
the same port as originally selected.

Change-Id: I0380b23daf57f1af04bf9a088779e7a43a3d1fd8
Reviewed-on: http://review.couchbase.org/53540
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>