MB-15782: Allow random key command to be run by clients
We plan on releasing a query tool along side the 4.0 release and
this tool requires using the random key command to implement the
describe feature of the tool. This command currently requires the
sender to use the admin password in order to use it. This change
allows the command to be used by all clients.

Change-Id: I169a090472ce3f1d18057f75697ceb8914a71cfd
Reviewed-on: http://review.couchbase.org/53561
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>