Topkeys: Restore 8 shards to implementation
Prior to the recent C++ refactor of topkeys[1] the top keys stored
were shared into 8 shards. This both increased the number of keys
actually stored (8 * MEMCACHED_TOP_KEYS) and allowed some amount of
concurrent access as each shard has it's own lock.

While the sharding may not be as necessary for concurrency as it once
was, reducing the number of keys stored down to 10 is a functional
regression. As such restore the sharding "as-was" for now.

[1]: 922a8d58eacd5775a1d94ef36bd8614f6c8bf379

Change-Id: I40e17971ff4393f5171fa3d8a1f968492e799c43
Reviewed-on: http://review.couchbase.org/53679
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>