MB-15910: Fix potential segfaults on WIN32 due to incorrect log format strings
Fix mismatch between LOG() formatting strings and the arguments. The
common pattern here is format strings which:

1) contain a char* (%s) formatter
2) preceded by a mis-sized integer formatter.

This results in printf() on 32bit Windows reading too much (or too
little) from the stack for the integer argument; meaning that when it
attempts to handle the following %s formatter and read the char* it
dereferences an invalid pointer.

Change-Id: I9e8bcddc2afc2f920d1db892da37cd588378e2a1
Reviewed-on: http://review.couchbase.org/53924
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>