Subdoc: Move document decompress to subdoc_fetch
Refactor when the subject document is obtained from the engine (and
decompressed if necessary) to the fetch stage of the execution, from
the operate phase.

This is a more 'correct' phase to perform this operation; and makes it
easier to add support for multi-path operations; which will only issue
subdoc_fetch() once, but subdoc_operate() multiple times.

Change-Id: I5c96745cdefb7e665f83839d83653925b6c500f3
Reviewed-on: http://review.couchbase.org/54177
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: buildbot <build@couchbase.com>