Subdoc: Convert traits from classes to objects
Sub-document API uses traits classes and templated functions to handle
the different subdoc commands. While this works well for single-path
operations, it is difficult to use for multi-path; where at runtime we
must dispatch a series of different commands based on the user
request. (Note runtime dispatch is also required for single-path, but
as there's only a one level of it needed it was reasonable to use the
subdoc_XXX_executor functions to perform the dispatch).

Replace the essentially stateless traits classes with actual objects;
with a series of get_traits<> functions (one per subdoc cmd) returning
a traits object for that command. Store the traits in the CmdContext,
and refactor the subdoc_executor() function and friends to use the
properties of the traits instead of template parameters.

Change-Id: I9bff25edb9d8e5acd37e126d093059640cfa4b63
Reviewed-on: http://review.couchbase.org/54179
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>