ActiveStream: Missing streamMutex locks
We are not acquiring the streamMutex lock in the
ActiveStream constructor method.  The mutex is
required because it makes potentially makes a call
to endStream, which in turn can call pushToReadyQ.

We are also not aqcuiring the lock in the function
incrBackfillRemaining, which is called from
DCPBackfill::create.  Given that it is just a stat
and does not need to stay consistent with other
variables instead of acquiring the streamMutex
lock, the backfillRemaining variable can be made
an atomic.

We are also not acquiring lock in the function
getReadyQueueMemory.  Given that it is just a read
instead of acquiring the streamMutex lock, the
readyQueueMemory variable can be made an atomic.

Change-Id: I5da7f986c4f939f28b258eabf0a951dfa7b60cc4
Reviewed-on: http://review.couchbase.org/54314
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Manu Dhundi <manu@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>