Add cbtrace
cbtrace is a packet trace program for the memcached binary
protocol. It reads a pcap file and by default dumps a histogram
of command timings. It was originally built in order to look
into some slow response times, but may be usable for others in
the future. Thats why the code is in the source repository.

NOTE: This program was hacked up in order to solve a single
problem, and I would be surprised if it didn't crash if you
try it out ;-)

Change-Id: I1ef86538971ca6e1de484c3c379ad34caf7564d6
Reviewed-on: http://review.couchbase.org/54380
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>