MB-13645: Subdoc MULTI_MUTATION validation
Implment support for validating sub-document MULTI_MUTATION commands.

- Add new traits for MULTI_MUTATION command.
- Enhance subdoc_multi_lookup_validator() to support both lookups and
  mutations, rename to subdoc_multi_validator.
- Add support for encoding MULTI_MUTATION commands to subdoc_encoders.
- Add testcases to memcached_mcbp_test for multi-path mutations.

Note: This patch changes the error-code returned when too many paths
are included in a multi-path request, from ERANGE to the
subdoc-specific error code SUBDOC_INVALID_COMBO. This reduces any
ambiguity over if the failure was due to the path count, or due to an
incorrect COUNTER operation.

Change-Id: I54ff9352007e68a2cdd0dd0c0e3218b49f43b616
Reviewed-on: http://review.couchbase.org/54415
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>