Create event test program
Memcached is using libevent and quite often we want to experiment
with some features on how libevent works in different scenarios
(the last time is if we do get a notification if the other side
close the socket while we've only got EV_WRITE registered). To
avoid having to whip up a complete new test program every time,
lets just commit the test program and have something to modify
the next time we need to run some experiments.

Change-Id: I9535ba406baa3242649e2bb0468bdc3c684fac24
Reviewed-on: http://review.couchbase.org/54447
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>