MB-16056: Replace uses of cb_assert() for invalid args with non-fatal error.
Where cb_assert() is used to assert that arguments to a
method/function are valid, replace with a non-fatal error - one of:

a). Throw std::invalid_argument, and catching at the caller(s).
b). Handle the error locally, logging where appropriate.
c). Refactor code so argument check is made compile-time.

Change-Id: I9fa4ede3f965459ebd12a7f08d1557658a9bbf76
Reviewed-on: http://review.couchbase.org/54903
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: buildbot <build@couchbase.com>