Minor tweaks to work better with clion
Opening memcached.cc in clion gives some red lines under
some statemens. This patch just fix those errors to make it
easier to work with clion.

Clion doesn't seem to like it when the variable name is the same
as the datatype.

Some variables hide a variable in the outer scope

Change-Id: Iefdc255c0967274dca939b2cbf44a79283d0e409
Reviewed-on: http://review.couchbase.org/54969
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>