afl-fuzz support.
1) Enable listening on stdin (writing back to stdout).

"stdin_listen" : true

This makes memcached listen to stdin (as well as the defined ports).
Development only feature as it forces libevent to avoid epoll.

2) Enable exit on connection close.

"exit_on_connection_close" : true

This config switch means that when any connection close memcached
will exit(0). This is required for afl-fuzz as it expects the
process to exit cleanly when a test is complete.

So when memcached reads 0 bytes (EOF) from stdin it will close
the connection and in turn exit(0).

afl-fuzz knows then to start a new test.

Change-Id: Ifd8ba62e0938fcfd8c3799ffda61e7ed1cada4e5
Reviewed-on: http://review.couchbase.org/55862
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: buildbot <build@couchbase.com>