Use CB_C,CXX_FLAGS_NO_OPTIMIZE for building unoptimized targets
Previously for targets which we wanted to always build without
optimization (e.g. unit tests for compilation speed reasons) we would
forcefully use the 'Debug' flags for those targets. However to do this
it was necessary to modify the Debug build flags, for example to
change the Windows CRT type to non-Debug so we didn't have mismatches
between different objects in the same binary.

This is undesirable as it means we cannot do true Debug builds on
Windows which use the Debug CRT.

Therefore switch to using the new CB_C,CXX_FLAGS_NO_OPTIMIZE
variable(s) which only disable optimzation, and don't affect C runtime
or similar. This will allow us (in a follow-up patch to remove our
overrides of CMAKE_{C,CXX}_FLAGS_DEBUG.

Change-Id: I0cdb30dea0026ddcdacadd6f8aceb48d40290abd
Reviewed-on: http://review.couchbase.org/55466
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>