MB-16056: Remove unsafe wildcard exception catch blocks in ExecutorThread
ExecutorThread has two sets of catch blocks where it catches both
std::exception and wildcard (...) and simply prints a warning to the
log and continues - this is not exception *handling*.

If ExecutorThread cannot take appropriate remedial action for an
exception (which in general is impossible for a catch-all exception)
then it shouldn't catch them at all, and rely on some higher-level
handler (or more likely the C++ runtime terminating the application
due to an unhandled exception.

Therefore remove these invalid catch blocks.

Change-Id: I1472914070008a16462ba41e5cd47c9eadb3a255
Reviewed-on: http://review.couchbase.org/55487
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>