MB-16056: Use correctly-sized type for VBucket identifiers
vBuckets are identified with a 16-bit value; however in VbucketMap it
uses various types (int and size_t) for vbucket identifiers.

This means that various places in the code need to perform range
checks to determine the values returned from VBucketMap are in range.

Simplify this by using a correctly-sized type for these methods - a
newly-created VBucket::id_type (typedef of uint16_t). This makes the
code safer (and simpler).

Static type-safety++

Change-Id: Ia496357ca74da6545b847b94a58976902dc4ca3b
Reviewed-on: http://review.couchbase.org/55491
Reviewed-by: abhinav dangeti <abhinav@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>