MB-16056: Use exceptions instead of cb_assert() [10]
Tenth set of changes to migrate away from cb_assert() and use
exceptions where applicable for invalid input / logic errors.

This is a "tidy-up" patch - fixes the remaining 13 or so cb_asserts
which I skipped over in the first pass as they had some form of subtly
to them. Highlights of interesting asserts:

- BgFetcher::start/stop, ConnNotifier::start/stop : These were
  asserting on taskId being >0. However taskId is set in the ExTask
  constructor using a atomic increment so cannot see any need for this
  assert.

- FailoverTable : These were postcondition checks that we had at least
  1 failover entry after various operations. However we can change
  them into precondiion checks on the input, and throw an exception if
  the proposed new value is not valid (i.e. turn it into a
  precondition check).

Change-Id: Iac6fff0b42e55e62fa6d0eeca446f8bfb094f3aa
Reviewed-on: http://review.couchbase.org/55608
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>