MB-16356: Subdoc: Add support for setting expiry for mutation commands
Add suport for setting expiration (TTL) of a document as part of
subdocument mutation commands, both single and multi-path.

The semantics of expiration is the same as fulldoc commands - i.e. a
(subdoc) mutation will always replace any existing expiration with a
new expiratio value. However, the encoding of the expiry has been
optimized (compared to fulldoc):

a) If an expiry is included in the extras, then that value is used as
  the new expiration if the mutation succeeds.

b) If an expiry isn't include in the extras, then an expiry value of
  '0' (i.e. never expire) is used.

These two cases are distinguished by the length of extras:

- For single-path extlen can be 3 bytes (expiration assumed to be 0)
  or 7 bytes.
- For multi-path (which previously didn't use extras) extlen can be 0
  bytes (expiration assumed to be 0) or 4 bytes.

Note there is _no_ concept of expiring only part of a document,
expiration continues to be a whole docuemnt concept.

Change-Id: Ia1e1c3d32762c57aebd3bb368862c042426678b8
Reviewed-on: http://review.couchbase.org/55564
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>