MB-16338: Increase test_defragmenter mapped memory thresholds
We are still seeing occasional failures in test_defragmenter - it
appears the recent change to the expected level of mapped memory was
too aggressive - recent failure:

    13876 Test failed: `current_mapped memory (which is 12865536) is lower than 90% of previous mapped (which is 13406208). ' (current_mapped >= 1.0 * double(previous_mapped))

(Note the "is lower than 90%" is a misnomer - the check is actually
against '100%' of the previous).

Changing the check back to <90% to give some additional error margin.

Change-Id: I61f76d29250f9054b680daf4a73842511e5b2288
Reviewed-on: http://review.couchbase.org/55593
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung@couchbase.com>