MB-12352 Remove unnesscary couchstore open/reads from backfill
CPU profiling of memcached during a rebalance showed that we opened and read
from each vbucket file twice. The source of the opens and reads were tracked
down to two places.

1. open/read - How many items for stat reporting using getNumItems
2. open/read - Scan for items using sequence number

To reduce the number of open/close we can integrate the getNumItems part
of this path into initScanContext.

The initScanContext call of CouchStore/ForestDB now collects the number
of items for the requested seqno delta. This reduces the opens/closes
on the backfill path.

Change-Id: I670f21047c1b91930294e6669d5809bfdb8bab83
Reviewed-on: http://review.couchbase.org/56156
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: buildbot <build@couchbase.com>