MB-16612: Move modules to "normal" locations
Previously we had bucket_engine load the modules on demand and
had ns_server provide the absolute path to the modules. Now we've
refactored the code to dynamically just load the modules during
startup (and may report link problems as part of the startup, and
not at a random time in the future (which may be weeks away)).

Change-Id: I001ce9cbc8f18f2aa466fb86c5832d5154cdfc38
Reviewed-on: http://review.couchbase.org/56314
Tested-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>