Allow BlockTimer to report "slow" blocks
Enhance BlockTimer to allow reporting to the log file blocks which
take longer than a given threshold to execute.

BlockTimer is now implemented in terms of GenericBlockTimer, a
template class which has a template parameter to specify that blocks
taking longer than THRESHOLD_MS should be logged. Note this mode isn't
enabled by default to minimise the overhead of BlockTimer (and also to
prevent log spam).

To enable for a single BlockTimer, simply change the type of your
BlockTimer variable to something like:

    GenericBlockTimer<100> timer(&histo, "block name");

where '100' indicates that if the block takes longer than 100 ms then
log it. Note the block name is required for the block to be logger.

To enable globally for *all* BlockTimers, simply change the definition
of BlockTimer (in histo.h) to:

typedef GenericBlockTimer<200> BlockTimer;

This will cause *all* block timers which take longer than 200
milliseconds to be logged (again, assuming they have a name
configured).

Change-Id: Ib6fa509e1f1c8f447d88f803d4474fa78bf257c3
Reviewed-on: http://review.couchbase.org/56391
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>