MB-16621: Ensure consumer thread terminates
If memcached is not shutdown gracefully it is
possible for ~Audit() to be invoked without
first calling shutdown_auditdaemon.  This means
the consumer thread will still exist.

The consumer thread blocks on the condition
variable events_arrived waiting for events.
Therefore when ~Audit() invokes cb_cond_destroy
it fails because the condition variable is busy.

This patch moves the termination of the consumer
thread to the clean_up function which is invoked
prior to cb_cond_destroy in the destructor.

Change-Id: I731d81fa747079678b137242a06c739477efd04c
Reviewed-on: http://review.couchbase.org/56565
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>