Added fatal_error function
Currently in most cases calls to exit are immediately
preceeded by a LOG_WARNING to report in the log why
the program is exiting.

However not all calls to exit are immediately preceded
by a call to LOG_WARNING.  This increases the risk of
exiting without reporting it in the log.

Therefore a new function "fatal_error" is being added
that combines the reporting in the log with the task
of exiting.  By using this function instead of directly
calling exit it is possible to ensure that the program
does not exit without first reporting the event in the
log.

Change-Id: I638f0b31481460e4c8cbc4d43d3d681314531932
Reviewed-on: http://review.couchbase.org/56648
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>