TopKeys: Optimize tracking of top keys - 4x speedup.
During some recent profiling it was noticed that top key tracking was
taking ~4.2% of all memcached process runtime. This is a large
overhead for a statistic operation.

Optimize the implementation of TopKeys. Full details in the code
comments, but in essence remove the majority of dynamic memory
allocation for the TopKeys tracking structure itself, by using a
std::vector<> for key statistics, and re-using a 'victim' slot when we
need to add a new key (and evict an old key) from the TopKey list.

After this change update topkeys is below 1% of the whole-process
profile (cbc-pillowfight -t 4 -m 250 -M 250 -U couchbase://localhost:12000/default -r 75 -I 40000).

Measuring using memcached_topkeys_bench:

Before:

    [ RUN      ] TopKeysTest.Basic
    [       OK ] TopKeysTest.Basic (1130 ms)

After:

    [ RUN      ] TopKeysTest.Basic
    [       OK ] TopKeysTest.Basic (267 ms)

Giving over a 4x speedup.

Change-Id: Ic9d38ed9139f6c4f756b6b34a562d1e9339ab185
Reviewed-on: http://review.couchbase.org/57045
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: buildbot <build@couchbase.com>