MB-16827: Create bloom filters for vbuckets upon bucket creation
Set up bloom filters for vbuckets immediately as opposed to waiting
for the first compaction to run for the following scenarios:
- upon creation of the bucket
- during rebalance

During warmup, the bloom filters will not be set up as we're not
aware of the deleted items and/or non-resident items at the moment.
The first compaction following the warmup will still be responsible
for setting up the bloom filters.

Change-Id: I40881a962b60e7c2c2eab497a2ec663cef293985
Reviewed-on: http://review.couchbase.org/57077
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: buildbot <build@couchbase.com>