MB-16984: Use Histogram/BlockTimer from platform
ep-engine has a histo.h file which contains code to measure the
duration of code blocks, and to record that information into a
histogram. To allow this code to be used by other projects it has been
moved into platform.

Update ep-engine to use the new, common version in platform and remove
ep-engine's own copy.

Change-Id: Idcb02d40167b2854422afdbdb80b13c3c4e266eb
Reviewed-on: http://review.couchbase.org/57589
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>