MB-16984 [Subdoc]: Add histogram of subjson exec times
Add a new TimingHistogram for how long subjson takes to parse the JSON
and execute a sub-document request. For multi-path requests the time
to execute /all/ paths is recorded.

The times for this are exposed via a new statistic:
`subjson_operation_times`. This returns a histogram of timings in JSON
format. mctimings has also been enhanced to be able to pretty-print
this statistic (it shares the same JSON format as the command timings.

Example usage:

    $ ./mctimings -h localhost:12000 -v subdoc_json_parse
    The following data is collected for "subdoc_json_parse"
    [   0 -  999]ns |######################################## - 67007316
    [   0 -    9]us | - 3
    Total: 67007319 operations

Change-Id: I879014c2c4e6fcf48530bf1c02c2bc8ad43896db
Reviewed-on: http://review.couchbase.org/57537
Well-Formed: buildbot <build@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>