MB-17027: Fix emfile_test hang due to too small log buffer
The emfile_test was hanging on certain configurations because the
formatted string it attempted to log was larger than the test buffer
size - 102 vs 100 bytes.

Fix by increasing the buffer size to 150 bytes.

(Note: the test deliberately sets a very small buffer size as to cause
a log rotation attempt, hence setting it to 100 bytes. However it
appears that if the environment the test was running in had a long
timezone string (e.g. '-08:00') then that was sufficient to create a
string longer than 100 bytes.)

Change-Id: I1b3db91b3208be82b117e1bb5ff0fce54bc65854
Reviewed-on: http://review.couchbase.org/57684
Well-Formed: buildbot <build@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>