MB-16745: Don't wait for the entire message to be sent
If the receiving end of the socket isn't draining the
data its receive-buffer may get filled and hence
blocking the bucket from being deleted.

Given that we want to send a response for a shutdown
request (and want to avoid race conditions for when the
bucket is starting to go away) we try to send the
response message on the socket _before_ checking if
the bucket is dying.

Change-Id: I98bd0128565ac23a1e82cf47c2df3ae83c9bc838
Reviewed-on: http://review.couchbase.org/57729
Well-Formed: buildbot <build@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>