mcset.c: Call to 'malloc' has an allocation size of 0 bytes
As identified by Clang static analyzer:

    File:	 programs/mcset/mcset.c
    Location:    line 47, column 19
    Description: Call to 'malloc' has an allocation size of 0 bytes

Undefined allocation of 0 bytes (CERT MEM04-C; CWE-131)

Note in this particular instance the allocation size of 0 bytes is
safe, as we never subsequently dereference the pointer if size is
zero. However to keep the code explicit and make Clang happy add an
explicit check for 0 before allocating memory.

Change-Id: I8b4542fafa1ce53ef25e7477ffc7ff5a24027922
Reviewed-on: http://review.couchbase.org/57915
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: buildbot <build@couchbase.com>