MB-16339: Ensure server is closed down cleanly in audit tests
The audit tests require the server to be shutdown to ensure
the audit events are written to disk (and so can be checked).

Previously the server was shutdown using stop_memcache_server().
On WIN32 this calls TerminateProcess(), which does not shutdown
the server cleanly and so the audit events might not be written
to disk.  The new approach is to send the shutdown command to
the server (PROTOCOL_BINARY_CMD_SHUTDOWN) which will cleanly
shutdown the server on all platforms.

This change now ensures that the audit tests pass and so are
re-enabled.

Change-Id: I3949de7e9981b22f8a9c5e2ba4ec2d3419ca1641
Reviewed-on: http://review.couchbase.org/58213
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>