Workaround for the thread-test race condition
There is currently a race condition during startup and
shutdown of the Thread class. We need to extend the
thread start and stop synchronization. This test fails
sporadically due to the fact that it keeps on
starting and stopping 4 worker threads, but only
use one of them.. Just change the test to use
one thread and we'll have to wait until that thread
is properly started before we can shut down.

Change-Id: Ibabc7323c520ebdaee113d6b039b32aca0c0b8e5
Reviewed-on: http://review.couchbase.org/58427
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>